renamed class and moved to auth package
